[#612] Merge validation code paths
Problem: Two code paths for validation were done by mistake. This commit
merges everything into one file.

Solution: Compare both code paths and merge them.
